% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{calc_direction_change}
\alias{calc_direction_change}
\title{Calculates difference in direction between two angles}
\usage{
calc_direction_change(oldDir, newDir)
}
\arguments{
\item{oldDir}{a number in the range of \[0, 360\)}

\item{newDir}{a number in the range of \[0, 360\)}
}
\value{
The difference between two angles. A numeric value between -180 and 180.
}
\description{
A helper function used by `short_turn\(\)` and `long_turn\(\)`
to determine the angular difference in degrees between two directions.
}
\examples{
calc_direction_change(20, 180) # => 160
calc_direction_change(350, 1)  # => 11
calc_direction_change(180, 0)  # => -180
calc_direction_change(180, 50) # => -130

}
\seealso{
short_turn, long_turn
}
