% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarrowPassesFilter.R
\name{id_narrow_pass}
\alias{id_narrow_pass}
\title{One Line Description}
\usage{
id_narrow_pass(data, remove = FALSE)
}
\arguments{
\item{data}{describe parameter}

\item{remove}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
id_narrow_pass(data)
id_narrow_pass(data, FALSE)
id_narrow_pass(data, TRUE)


}

