% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassEndFilter.R
\name{id_pass_end}
\alias{id_pass_end}
\title{Identifies turns.}
\usage{
id_pass_end(data, remove = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{remove}{a Boolean. Defaults to FALSE.}
}
\value{
A dataframe. Added column called cydr_Error. If a potential error
has been identified this column is TRUE. Otherwise it is FALSE.
}
\description{
id_pass_turns identifies turns at the end of passes within a
field.
}
\examples{
id_pass_end(field1)
id_pass_end(field1, remove=TRUE)
}

