% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassTurnFilter.R
\name{id_pass_turns}
\alias{id_pass_turns}
\title{Identifies possibily erroneous pass-turn data.}
\usage{
id_pass_turns(data, remove = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{remove}{a Boolean. Defaults to FALSE.}
}
\value{
A new dataframe with a column cydrError indicating whether an error
was identified.
}
\description{
Identifies possibily erroneous pass-turn data.
}
\examples{
id_pass_turns(field1)
id_pass_turns(field1, remove=TRUE)
}

