% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidualFilter.R
\name{id_residuals}
\alias{id_residuals}
\title{One Line Description}
\usage{
id_residuals(data, remove = FALSE, standard_devs = 2)
}
\arguments{
\item{data}{describe parameter}

\item{remove}{describe parameter}

\item{standard_devs}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
id_residuals(data)
id_residuals(data, FALSE)
id_residuals(data, TRUE)


}

