% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedFilter.R
\name{id_speed}
\alias{id_speed}
\title{One Line Description}
\usage{
id_speed(data, remove = FALSE, type = "both")
}
\arguments{
\item{data}{describe parameter}

\item{remove}{describe parameter}

\item{type}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
id_speed(data, type="both")
id_speed(data, FALSE, type="low")
id_speed(data, TRUE, type="high")


}

