% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StopPileFilter.R
\name{id_stop_piles}
\alias{id_stop_piles}
\title{One Line Description}
\usage{
id_stop_piles(data, remove = FALSE)
}
\arguments{
\item{data}{describe parameter}

\item{remove}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
id_stop_piles(data)
id_stop_piles(data, FALSE)
id_stop_piles(data, TRUE)


}
