% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarrowPassesFilter.R
\name{is_LongTurn}
\alias{is_LongTurn}
\title{One Line Description}
\usage{
is_LongTurn(lag, lead, threshold = 100)
}
\arguments{
\item{lag}{describe parameter}

\item{lead}{describe parameter}

\item{threshold}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
is_LongTurn(data)
}
\seealso{
is_ShortTurn
}
