% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarrowPassesFilter.R
\name{is_ShortTurn}
\alias{is_ShortTurn}
\title{One Line Description}
\usage{
is_ShortTurn(lag, lead, threshold = 45)
}
\arguments{
\item{lag}{describe parameter}

\item{lead}{describe parameter}

\item{threshold}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
is_ShortTurn(lead, lag, threshold)
}
\seealso{
is_LongTurn
}
