% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarrowPassesFilter.R
\name{narrow_pass}
\alias{narrow_pass}
\title{Identify Narrow Passes in a dataframe}
\usage{
narrow_pass(data, remove = FALSE, passes = "cydr_PassNum")
}
\arguments{
\item{data}{a dataframe.}

\item{remove}{a boolean. Defaults to FALSE}

\item{passes}{a string, either 'cydr' or the column in the data which specifies
an observations pass number. Defaults to 'cydr'.}
}
\value{
A dataframe. Added column called cydr_Narrow_Pass. If a point's pass has
been identified as narrow this column is TRUE. Otherwise it is FALSE.
}
\description{
Identify narrow passes within a data frame representing a field's
yield. It creates a new data frame with an additional column to flag narrow
passes. This is particularily useful in conjunction with the other cydr 
identification functions to identify erroneous observations in agricultural 
yield data.
}
\examples{
narrow_pass(data)
narrow_pass(data, FALSE)
narrow_pass(data, TRUE)


}

