% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarrowPasses.R
\name{narrow_passes}
\alias{narrow_passes}
\title{Identify narrow passes}
\usage{
narrow_passes(data, remove=FALSE, passColumn=NULL, diff=0.15)
}
\arguments{
\item{data}{a dataframe standardized and outputted from AgLeader.}

\item{remove}{a boolean. Defaults to \code{FALSE}. Indicates whether to remove
identified errors.}

\item{passColumn}{a string or \code{NULL}. Defaults to \code{NULL}. Indicates what
column specifies an observation's pass number. If \code{NULL} the built-in helper
function \code{number_passes()} will be used to assign pass numbers to each observation.}

\item{diff}{a numeric value between 0 and 1. Defaults to 0.15. Determines the allowable relative
difference between neighbouring passes.}
}
\value{
A dataframe with an added column called \code{cydr_NarrowPassError}. This column
will be set to \code{TRUE} if it meets the criteria for an erroneous observation.
}
\description{
Adds a column called \code{cydr_NarrowPassError} to a dataframe to 
identify observations associated with narrow passes. 

Will identify all observations within narrow passes as erroneous. A pass is deemed
narrow if it has an average yield less than 1-\code{diff} times that of either its
neighbours.
}
\examples{
narrow_passes(data)
narrow_passes(data, remove=TRUE, diff=0.25) # Allows differences of up to 25\%
narrow_passes(data, passColumn="Pass_Num") # Will use Pass_Num to number passes as th

}
\seealso{
Other core functions: \code{\link{pass_end_turns}},
  \code{\link{speed}}
}
