% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarrowPassesFilter.R
\name{number_passes}
\alias{number_passes}
\title{Determines pass numbers in a dataframe}
\usage{
number_passes(data)
}
\arguments{
\item{data}{describe parameter}
}
\value{
a new dataframe containing a new column called cydr_PassNum 
representing
}
\description{
Determines pass numbers in a dataframe
}
\examples{
number_passes(data)


}

