% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarrowPassesFilter.R
\name{plot_cydr_Passes}
\alias{plot_cydr_Passes}
\title{One Line Description}
\usage{
plot_cydr_Passes(data)
}
\arguments{
\item{lag}{describe parameter}

\item{lead}{describe parameter}

\item{threshold}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
update_progress_bar(pos, len)

}
