% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidualOutliers.R
\name{residual_outliers}
\alias{residual_outliers}
\title{Identify outlying yields}
\usage{
residual_outliers(data, remove=FALSE, type="both", sd=2)
}
\arguments{
\item{data}{a dataframe, standardized and outputted from AgLeader.}

\item{remove}{a boolean. Defaults to \code{FALSE}. Indicates whether to remove
identified errors.}

\item{type}{one of \code{"high"}, \code{"low"}, or \code{"both"}. Indicates
which types of data to identify as erroneous \code{"high"} will identify, 
high yields, \code{"low"} will identify low speeds, and \code{"both"} will
identify high and low speeds.}

\item{sd}{a number >= 0. Defaults to 2. Used as the standard deviation 
threshold for error identification.}
}
\value{
A dataframe with an added column called \code{cydr_ResidualError}. 
This column will be set to \code{TRUE} if an observation is deemed erroneous.

If \code{remove = TRUE} all observations cydr identifies as erroneous are
removed from the returned dataframe.
}
\description{
Adds a column called \code{cydr_ResidualError} to a 
dataframe to identify observations associated with outlying yields. 

Will identify all observations with yields greater than \code{sd} standard
deviations away from the mean. If \code{remove} is \code{TRUE} all 
observations associated with a residual error will be removed from the 
dataframe.
}
\examples{
residual_outliers(data)
residual_outliers(data, TRUE)
residual_outliers(data, TRUE, type="low", sd=3)

}
\seealso{
Other core functions: \code{\link{narrow_passes}},
  \code{\link{pass_end_turns}}, \code{\link{speed}}
}
