% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedFilter.R
\name{speed}
\alias{speed}
\title{Identify outlying speeds.}
\usage{
speed(data, ...)
}
\arguments{
\item{data}{a dataframe standardized and outputted from AgLeader}

\item{remove}{a boolean. Defaults to \code{FALSE}.}

\item{type}{one of \code{"high"}, \code{"low"}, or \code{"both"} to determine 
which types of data to remove. \code{"high"} will identify fast speeds, 
\code{"low"} will identify slow speeds, and \code{"both"} will identify data 
associated with either fast or slow speeds.}

\item{sd}{a number >= 0. Defaults to 2. Used as the standard deviation 
threshold for error identification.}
}
\value{
a dataframe with an added column called \code{cydr_SpeedError}. This column
will be set to \code{TRUE} if it meets the criteria for an erroneous observation.

If \code{remove = TRUE} all observations cydr identifies as erroneous are 
removed from the returned dataframe.
}
\description{
Adds a column called \code{cydr_SpeedError} to a dataframe to identify
observations associated with outlying speeds. 

Will identify all observations with speeds greater than \code{sd} standard
deviations from the mean. If \code{remove} is \code{TRUE} all erroneous 
observations will be removed from the dataframe.
}
\examples{
speed(data)
speed(data, type="both")
speed(data, FALSE, type="low")
speed(data, TRUE, type="high")

}

