% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToBeNamed.R
\name{summarize_cydrErrors}
\alias{summarize_cydrErrors}
\title{One line description}
\usage{
summarize_cydrErrors(data)
}
\arguments{
\item{param1}{a dataframe}

\item{param2}{a boolean. Defaults to FALSE.}
}
\value{
Value
}
\description{
description
}
\examples{
id_pass_end(field1)
id_pass_end(field1, remove=TRUE)
}

