% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NarrowPassesFilter.R
\name{update_progress_bar}
\alias{update_progress_bar}
\title{One Line Description}
\usage{
update_progress_bar(pos, len)
}
\arguments{
\item{lag}{describe parameter}

\item{lead}{describe parameter}

\item{threshold}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
update_progress_bar(pos, len)

}

