% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassEndFilter.R
\name{calc_direction_change}
\alias{calc_direction_change}
\title{A helper function for ____ which calculates difference between two angles.}
\usage{
calc_direction_change(oldDir, newDir)
}
\arguments{
\item{oldDir}{a number in the range of \[0, 360\)}

\item{newDir}{a number in the range of \[0, 360\)}
}
\value{
The difference between two angles.
}
\description{
A helper function for ____ which calculates difference between two angles.
}
\examples{
calc_direction_change(20, 180)
calc_direction_change(350, 1)
calc_direction_change(0, 180)

}
