% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidualFilter.R
\name{id_residuals}
\alias{id_residuals}
\title{Identifies data points whose Yld_Vol_Dr value is more than a specified
number of deviations from the mean.}
\usage{
id_residuals(data, remove = FALSE, num_sd = 2)
}
\arguments{
\item{data}{a dataframe.}

\item{remove}{a boolean. Defaults to FALSE.}

\item{num_sd}{a integer in the range of [1, 4]. Defaults to 2.}
}
\value{
A dataframe. Added column called cydr_ResidualErrors. If a potential error
has been identified this column is TRUE. Otherwise it is FALSE.
If remove == TRUE, any observation whose cydr_ResidualErrors attribute is TRUE is
removed.
}
\description{
Identifies data points whose Yld_Vol_Dr value is more than a specified
number of deviations from the mean.
}
\examples{
id_residuals(data)
id_residuals(data, FALSE)
id_residuals(data, TRUE)
}

