% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{long_turn}
\alias{long_turn}
\title{One Line Description}
\usage{
long_turn(lag, lead, threshold = 100)
}
\arguments{
\item{lag}{describe parameter}

\item{lead}{describe parameter}

\item{threshold}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
long_turn(data)
}
\seealso{
short_turn
}
