% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{number_passes}
\alias{number_passes}
\title{Groups a yield data frame into passes}
\usage{
number_passes(data)
}
\arguments{
\item{data}{a data frame representing a set of yield data. The 
data frame must include a `Track_deg_` column.}
}
\value{
a data frame with an added column called `cydr_PassNum`. 
This column stores the pass number of each observation.
}
\description{
a helper function used by `narrow_passes` to group
observations into passes. Used if no `passColumn` is supplied to the
`narrow_passes` function. 

 Will loop through the dataset and assign pass numbers to each 
 observation.
}
\seealso{
narrow_passes
}
