% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{short_turn}
\alias{short_turn}
\title{One Line Description}
\usage{
short_turn(lag, lead, threshold = 45)
}
\arguments{
\item{lag}{describe parameter}

\item{lead}{describe parameter}

\item{threshold}{describe parameter}
}
\value{
value what does it return?
}
\description{
One Line Description
}
\examples{
short_turn(lead, lag, threshold)
}
\seealso{
long_turn
}
